requirements.txt
pyarr==5.2.0
PyYAML==6.0.1
requests==2.31.0
yt-dlp==2024.3.10
